routes -> refact: isolate useMatch() custom hook
