+++ packages/site-admin/src/router/hooks/index.ts
@@ -1,1 +1,2 @@
 export { default as useLocation } from './use-location';
+export { default as useMatch } from './use-match';
