+++ packages/site-admin/src/router/types.ts
@@ -18,3 +18,14 @@ export interface Match {
 	query?: Record< string, any >;
 	params?: Record< string, any >;
 }
+
+export type LocationWithQuery = Location & {
+	query?: Record< string, any >;
+};
+
+export interface Route {
+	name: string;
+	path: string;
+	areas: Record< string, ReactNode >;
+	widths: Record< string, number >;
+}
